+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ability for user to provide a preference value to the `COCOSO` ranking method. This value preferences which kind of scoring strategy the `COCOSO` ranking method uses.
 - Ability for user to provide a preference or `None` to the `ERVD` and `WASPAS` ranking methods. The the user provides `None`, the ranking methods default to the recommended default values.
 - CILOS weighting method. Adding this also required adding an additional error variant, `ValidationError::SingularMatrix`, in the event a certain matrix isn't a square matrix when it actually should be.
+- IDOCRIW weighting method.
 
 ### Changed